feat: Client can see current and history reservations

+++ demo/src/main/java/app/domain/CottageReservation.java
@@ -26,7 +26,7 @@ public class CottageReservation {
     private LocalDateTime endTime;
 
     @Column
-    private double price;
+    private Integer price;
 
     @ManyToOne(fetch = FetchType.EAGER)
     @JoinColumn(name = "client_id")
@@ -40,7 +40,7 @@ public class CottageReservation {
 
     public CottageReservation(){}
 
-    public CottageReservation(LocalDateTime startTime, LocalDateTime endTime,double price, Client client, Cottage cottage) {
+    public CottageReservation(LocalDateTime startTime, LocalDateTime endTime,Integer price, Client client, Cottage cottage) {
         this.startTime = startTime;
         this.endTime = endTime;
         this.price = price;
@@ -48,7 +48,7 @@ public class CottageReservation {
         this.cottage = cottage;
     }
 
-    public CottageReservation(LocalDateTime startTime, double price, Client client, Cottage cottage) {
+    public CottageReservation(LocalDateTime startTime, Integer price, Client client, Cottage cottage) {
         this.startTime = startTime;
         this.endTime = startTime.plusDays(1);
         this.price = price;
@@ -80,9 +80,9 @@ public class CottageReservation {
         this.endTime = endTime;
     }
 
-    public double getPrice() { return price; }
+    public Integer getPrice() { return price; }
 
-    public void setPrice(double price) { this.price = price; }
+    public void setPrice(Integer price) { this.price = price; }
 
     public Client getClient() { return client; }
