+++ demo/src/main/java/app/dto/CottageReservationDTO.java
@@ -16,13 +16,13 @@ public class CottageReservationDTO {
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm", shape = JsonFormat.Shape.STRING)
     private LocalDateTime endTime;
 
-    private double price;
+    private Integer price;
     private Integer clientId;
     private Integer cottageId;
 
     public CottageReservationDTO(){}
 
-    public CottageReservationDTO(LocalDateTime startTime, LocalDateTime endTime,double price, Integer clientId, Integer cottageId) {
+    public CottageReservationDTO(LocalDateTime startTime, LocalDateTime endTime,Integer price, Integer clientId, Integer cottageId) {
         this.startTime = startTime;
         this.endTime = endTime;
         this.price = price;
@@ -30,7 +30,7 @@ public class CottageReservationDTO {
         this.cottageId = cottageId;
     }
 
-    public CottageReservationDTO(LocalDateTime startTime, double price, Integer clientId, Integer cottageId) {
+    public CottageReservationDTO(LocalDateTime startTime, Integer price, Integer clientId, Integer cottageId) {
         this.startTime = startTime;
         this.endTime = startTime.plusDays(1);
         this.price = price;
@@ -42,9 +42,9 @@ public class CottageReservationDTO {
 
     public void setStartTime(LocalDateTime startTime) { this.startTime = startTime; }
 
-    public double getPrice() { return price; }
+    public Integer getPrice() { return price; }
 
-    public void setPrice(double price) { this.price = price; }
+    public void setPrice(Integer price) { this.price = price; }
 
     public Integer getClientId() { return clientId; }
